+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG (v2)
 
+## v2.1.0
+- VersioningHelper supports Composer versioning (1.0.0-alpha1 in stead of 1.0.0-alpha.1) 
+- Updated [VersionHelper documentation](docs/Helpers.md#bumpsemanticversionstring-currentversion-enumbumptype-type-enumbumpprerelease-prerelease--null-string-sequencesplitter---string)
+
 ## v2.0.0
 A complete rewrite of SEworqs Commons. Please check documentation on how to use the new and improved SEworqs Commons!